remove python 2 stuff from tests
