+++ .travis.yml
@@ -27,9 +27,9 @@ addons:
             - texlive-latex-extra
             - dvipng
 python:
-    - 2.7
     - 3.5
     - 3.6
+    - 3.7
 env:
     global:
         # The following versions are the 'default' for tests, unless
@@ -49,19 +49,19 @@ matrix:
     allow_failures:
         - os: osx
         - os: linux
-          python: 3.5
+          python: 3.6
           env: MAIN_CMD='pycodestyle' SETUP_CMD='--count py/desiutil'
 
     include:
         # Check for sphinx doc build warnings - we do this first because it
         # runs for a long time
         - os: linux
-          python: 3.5
+          python: 3.6
           env: SETUP_CMD='build_sphinx --warning-is-error'
 
         # Coverage test, pass the results to coveralls.
         - os: linux
-          python: 3.5
+          python: 3.6
           env: MAIN_CMD='coverage' SETUP_CMD='run --source=desiutil setup.py test'
 
         # Experimental OS X test.
@@ -73,7 +73,7 @@ matrix:
 
         # PEP 8 compliance.
         - os: linux
-          python: 3.5
+          python: 3.6
           env: MAIN_CMD='pycodestyle' SETUP_CMD='--count py/desiutil'
 
 # before_install:
