--- doc/rtd-requirements.txt
@@ -1,1 +1,0 @@
-setuptools
