+++ py/desiutil/test/__init__.py
@@ -1,12 +1,35 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 # -*- coding: utf-8 -*-
-from __future__ import absolute_import
-#
-# Don't import this function.  It causes a scary-seeming warning of the form:
-# .../lib/python3.5/runpy.py:125: RuntimeWarning:
-#     'desiutil.test.desiutil_test_suite' found in sys.modules after
-#     import of package 'desiutil.test', but prior to execution of
-#     'desiutil.test.desiutil_test_suite'; this may result in unpredictable behaviour
-#        warn(RuntimeWarning(msg))
-#
-# from .desiutil_test_suite import runtests
+"""
+desiutil.test
+=============
+
+Used to initialize the unit test framework via ``python setup.py test``.
+"""
+import unittest
+
+
+def desiutil_test_suite():
+    """Returns unittest.TestSuite of desiutil tests.
+
+    This is factored out separately from runtests() so that it can be used by
+    ``python setup.py test``.
+    """
+    from os.path import dirname
+    py_dir = dirname(dirname(__file__))
+    # print(desiutil_dir)
+    return unittest.defaultTestLoader.discover(py_dir,
+                                               top_level_dir=dirname(py_dir))
+
+
+def runtests():
+    """Run all tests in desiutil.test.test_*.
+    """
+    # Load all TestCase classes from desiutil/test/test_*.py
+    tests = desiutil_test_suite()
+    # Run them
+    unittest.TextTestRunner(verbosity=2).run(tests)
+
+
+if __name__ == "__main__":
+    runtests()
