+++ py/desiutil/test/test_depend.py
@@ -2,9 +2,6 @@
 # -*- coding: utf-8 -*-
 """Test desiutil.depend.
 """
-from __future__ import (absolute_import, division,
-                        print_function, unicode_literals)
-
 import unittest
 import sys
 from collections import OrderedDict
