+++ py/desiutil/test/test_plots.py
@@ -2,9 +2,6 @@
 # -*- coding: utf-8 -*-
 """Test desiutil.plots.
 """
-from __future__ import (absolute_import, division,
-                        print_function, unicode_literals)
-# The line above will help with 2to3 support.
 import unittest
 import shutil
 import tempfile
